Удали остатки state-model & state-controller
